app/requirements.txt
Flask==3.1.2
gunicorn
python-dotenv
pydantic
Flask-Limiter
PyJWT
cryptography
flask-cors
httpx
SQLAlchemy
psycopg2-binary
alembic
Flask-Caching
redis
python-json-logger
prometheus-client
